+++ src/modules/core.js
@@ -1,27 +1,21 @@
 // **render**
 //
-// Render a template literal inside a container element.
+// Render a component inside its container element.
 //
-Combo.render = function(el, value) {
+Combo.render = function(component, props) {
+	let el = document.getElementById(component.root);
+	//
+	// Remove all child elements from the container.
+	//
 	if(el.firstChild !== undefined) {
 		while(el.firstChild) {
 			el.removeChild(el.firstChild);
 		}
 	}
-
-	el.insertAdjacentHTML("beforeEnd", value);
+	//
+	// Render the component, passing over properties.
+	//
+	el.insertAdjacentHTML("beforeEnd", component.render(props));
 };
 
-// **map**
-//
-// Return a concatenated string from elements in an Array.
-//
-Combo.map = function(obj, fn) {
-	var result = "";
-
-	obj.map((item) => {
-		result += fn(item);	
-	});
-
-	return result;
-};
+